mcp_server_template/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-server-stub"
version = "0.1.0"
description = "Bootstrap Hydra + FastMCP server stub"
requires-python = ">=3.10"
dependencies = [
  "hydra-core>=1.3,<2.0",
  "mcp>=1.9.4,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4,<9.0",
]

[project.scripts]
mcp-server = "mcp_server.main:main"

[tool.pytest.ini_options]
addopts = "--import-mode=importlib"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"mcp_server.conf" = ["*.yaml"]
